pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "moodle-vault"
version = "0.1.0"
description = "Automated Moodle scraping tool for file downloading and course organization."
requires-python = ">=3.10"
authors = [
    { name = "Matías Leonel Zalazar", email = "matias.zalazar@icloud.com" },
]
dependencies = [
    "selenium>=4.10.0,<5.0.0",
    "requests>=2.31.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "typer>=0.12.0",
    "loguru>=0.7.2",
]

[project.scripts]
moodle = "cli.commands:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "flake8>=7.0.0",
    "isort>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["cli", "scripts"]
